+++ CHANGELOG.md
@@ -7,8 +7,8 @@
 * `ExportButton` will auto-disable itself if bound to an empty `GridModel`. This helper button will
   now also throw a console warning (to alert the developer) if `gridModel.enableExport != true`.
 
-* 'DateInput' supports a new 'enabledTextInput' prop. When this property is set to false, 'DateInput' 
-  will be entirely driven by the provided date picker. Additionally, 'DateInput' styles have been
+* `DateInput` supports a new `enableTextInput` prop. When this property is set to false, `DateInput` 
+  will be entirely driven by the provided date picker. Additionally, `DateInput` styles have been
   improved for its various modes to more clearly convey its functionality.